fix(ads): one Ads hub for every ad channel, fixing LinkedIn Ads "Manage"
Content -> LinkedIn Ads -> Manage bounced straight back to
/dashboard/integrations. Neither source of the deep-link knew where to
manage the channel: the CMS catalog row has no display.dashboardPath, and
the static channels.config row was a stale `coming_soon` with no path
either, so STATIC_DASHBOARD_PATHS never carried linkedin_ads. The hub fell
through to its "send them to the connect grid" branch - for a channel the
org had already connected.

Rather than add a page per ad platform, /dashboard/ads becomes the ONE ads
surface with a `?channel=` selector:

- ads-channels.ts - channel registry (label, providerKey, description,
  crons, invalidation keys) + resolveAdsChannel: explicit param wins, else
  first connected channel, else first registered.
- ads/page.tsx - hub shell owning the header, tab selector, and
  CronToolbar; panels are a total Record<AdsChannelKey, ...> so adding a
  channel without a panel is a compile error.
- ads/_components/linkedin-ads-panel.tsx - was ads/page.tsx.
- ads/_components/x-ads-panel.tsx - was content/x-ads/page.tsx; that route
  is now a redirect to /dashboard/ads?channel=x.
- channels.config.ts - linkedin_ads live + /dashboard/ads?channel=linkedin
  (also unblocks the api catalog seed, which derives display.dashboardPath
  from this file); x_ads repointed at the hub.

Also stop the silent bounce class of bug: resolveChannelCta now reports
manageUnavailable (connected, but no path in either source) and the row
renders a disabled "Connected" plus a dev console.error, instead of a
Manage button that dumps the user back on the connect grid.
